+++ src/py/litestar_vite/inertia/__init__.py
@@ -1,17 +1,17 @@
-from .config import InertiaConfig
-from .exception_handler import create_inertia_exception_response, exception_to_http_response
-from .middleware import InertiaMiddleware
-from .plugin import InertiaPlugin
-from .request import InertiaDetails, InertiaHeaders, InertiaRequest
-from .response import (
+from litestar_vite.inertia import helpers
+from litestar_vite.inertia.config import InertiaConfig
+from litestar_vite.inertia.exception_handler import create_inertia_exception_response, exception_to_http_response
+from litestar_vite.inertia.helpers import error, get_shared_props, js_routes_script, lazy, share
+from litestar_vite.inertia.middleware import InertiaMiddleware
+from litestar_vite.inertia.plugin import InertiaPlugin
+from litestar_vite.inertia.request import InertiaDetails, InertiaHeaders, InertiaRequest
+from litestar_vite.inertia.response import (
     InertiaBack,
     InertiaExternalRedirect,
     InertiaRedirect,
     InertiaResponse,
-    error,
-    get_shared_props,
-    share,
 )
+
 from .routes import generate_js_routes
 
 __all__ = (
@@ -30,5 +30,8 @@ __all__ = (
     "exception_to_http_response",
     "generate_js_routes",
     "get_shared_props",
+    "helpers",
+    "js_routes_script",
+    "lazy",
     "share",
 )
