+++ src/py/litestar_vite/inertia/plugin.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from anyio.from_thread import BlockingPortalProvider
 from litestar.plugins import InitPluginProtocol
 
 if TYPE_CHECKING:
@@ -23,7 +22,7 @@ def set_js_routes(app: Litestar) -> None:
 class InertiaPlugin(InitPluginProtocol):
     """Inertia plugin."""
 
-    __slots__ = ("config", "portal")
+    __slots__ = ("_portal", "config")
 
     def __init__(self, config: InertiaConfig) -> None:
         """Initialize ``Inertia``.
@@ -32,11 +31,6 @@ class InertiaPlugin(InitPluginProtocol):
             config: Inertia configuration.
         """
         self.config = config
-
-        self.portal = BlockingPortalProvider()
-
-    def get_portal(self) -> BlockingPortalProvider:
-        return self.portal
 
     def on_app_init(self, app_config: AppConfig) -> AppConfig:
         """Configure application for use with Vite.
@@ -52,9 +46,10 @@ class InertiaPlugin(InitPluginProtocol):
         from litestar.utils.predicates import is_class_and_subclass
 
         from litestar_vite.inertia.exception_handler import exception_to_http_response
+        from litestar_vite.inertia.helpers import DeferredProp, StaticProp
         from litestar_vite.inertia.middleware import InertiaMiddleware
         from litestar_vite.inertia.request import InertiaRequest
-        from litestar_vite.inertia.response import DeferredProp, InertiaBack, InertiaResponse, StaticProp
+        from litestar_vite.inertia.response import InertiaBack, InertiaResponse
 
         for mw in app_config.middleware:
             if isinstance(mw, DefineMiddleware) and is_class_and_subclass(
